Refactor and enhance blob storage service
- Changed `RangeEnd` type to `int64` for consistency across the system.
- Added `Size` field to `CompleteUploadResponse` and updated handlers to register blob size.
- Implemented `GetSize` method in `Blob` and updated its constructor to include `size`.
- Fully implemented `BlobExists` handler with tenant, project, and repository validation.
- Fixed API routes for starting blob uploads by appending trailing slashes.
- Refactored database schema and initialization logic in memory database.
- Adjusted setup and middleware to reflect these updates.

+++ internal/jsontypes/UploadSession.go
@@ -16,6 +16,6 @@ type UploadSession struct {
 	ProjectSlug    string         `json:"projectSlug"`
 	RepositorySlug string         `json:"repositorySlug"`
 	RepositoryId   uuid.UUID      `json:"repositoryId"`
-	RangeEnd       int            `json:"rangeEnd"`
+	RangeEnd       int64          `json:"rangeEnd"`
 	DigestState    []byte         `json:"digestState"`
 }
